Fix UriHandlingSpec
